+++ nex_base/lib/nex_base/repo/postgres.ex
@@ -4,7 +4,10 @@ defmodule NexBase.Repo.Postgres do
     adapter: Ecto.Adapters.Postgres
 
   def init(_type, config) do
-    repo_config = Application.get_env(:nex_base, :repo_config, [])
+    name = config[:name]
+    repo_config =
+      (name && Application.get_env(:nex_base, name, nil)) ||
+      Application.get_env(:nex_base, :repo_config, [])
     {:ok, Keyword.merge(config, repo_config)}
   end
 end