+++ nex_base/lib/nex_base/repo/sqlite.ex
@@ -5,7 +5,10 @@ if Code.ensure_loaded?(Ecto.Adapters.SQLite3) do
       adapter: Ecto.Adapters.SQLite3
 
     def init(_type, config) do
-      repo_config = Application.get_env(:nex_base, :repo_config, [])
+      name = config[:name]
+      repo_config =
+        (name && Application.get_env(:nex_base, name, nil)) ||
+        Application.get_env(:nex_base, :repo_config, [])
       {:ok, Keyword.merge(config, repo_config)}
     end
   end